Major redesign: move files between folders to achieve an easier to follow structure.:
