+++ src/httpserver/server/WebServer.java
@@ -1,8 +1,7 @@
 package httpserver.server;
 
-import httpserver.handlers.RequestRouter;
+import httpserver.request.RequestRouter;
 import httpserver.request.RequestParser;
-import httpserver.response.ResponseWriter;
 
 import java.io.*;
 import java.net.ServerSocket;
