+++ src/httpserver/utilities/Method.java
@@ -1,4 +1,4 @@
-package httpserver;
+package httpserver.utilities;
 
 public enum Method {
     GET, POST, HEAD, PUT, PATCH, DELETE, OPTIONS
