+++ test/requestTests/RequestTests.java
@@ -1,6 +1,6 @@
 package requestTests;
 
-import httpserver.Method;
+import httpserver.utilities.Method;
 import httpserver.request.Request;
 import org.junit.Before;
 import org.junit.Test;
