+++ test/serverTests/WebServerTests.java
@@ -1,6 +1,6 @@
 package serverTests;
 
-import httpserver.handlers.RequestRouter;
+import httpserver.request.RequestRouter;
 import httpserver.request.RequestParser;
 import httpserver.server.*;
 import httpserver.server.mocks.MockServerSocket;
